Move time window validation to server side
