+++ viz/metrics-api/grpc_server.go
@@ -15,6 +15,7 @@ import (
 	promv1 "github.com/prometheus/client_golang/api/prometheus/v1"
 	log "github.com/sirupsen/logrus"
 	"google.golang.org/grpc"
+	"gopkg.in/yaml.v2"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/labels"
 )
@@ -236,4 +237,45 @@ func (s *grpcServer) ListServices(ctx context.Context, req *pb.ListServicesReque
 	}
 
 	return &pb.ListServicesResponse{Services: svcs}, nil
+}
+
+// validateTimeWindow returns an error if the Prometheus scrape interval
+// is longer than the query time window. This is an opportunistic, best-effort
+// validation: if we cannot determine the Prometheus scrape interval for any
+// reason, we do not return an error.
+func (s *grpcServer) validateTimeWindow(ctx context.Context, window string) error {
+	config, err := s.prometheusAPI.Config(ctx)
+	if err != nil {
+		return nil
+	}
+
+	type PrometheusConfig struct {
+		Global map[string]string
+	}
+
+	var prom PrometheusConfig
+	err = yaml.Unmarshal([]byte(config.YAML), &prom)
+	if err != nil {
+		return nil
+	}
+
+	scrape_interval_str, found := prom.Global["scrape_interval"]
+	if !found {
+		return nil
+	}
+
+	scrape_interval, err := time.ParseDuration(scrape_interval_str)
+	if err != nil {
+		return nil
+	}
+
+	t, err := time.ParseDuration(window)
+	if err != nil {
+		return err
+	}
+
+	if t < scrape_interval {
+		return fmt.Errorf("Time window (%s) must be at least as long as the Prometheus scrape interval (%s)", window, scrape_interval)
+	}
+	return nil
 }
